+++ splash/tests/test_execute.py
@@ -3574,6 +3574,57 @@ class IsolationTest(BaseLuaRenderTest):
         self.assertEqual(resp.text, "value")
 
 
+class IndexedDBTest(BaseLuaRenderTest):
+    DETECT_INDEXEDDB_JS = """
+    function main(splash) {
+        if (!window.indexedDB) {
+            splash.resume("no indexed DB");
+            return;
+        }
+        var request = window.indexedDB.open("MyTestDatabase", 3);
+        request.onerror = function(event) {
+            splash.resume("can't open IndexedDB database");            
+        };
+        request.onsuccess = function(event) {
+            splash.resume("ok");
+        };                
+    }
+    """
+    def test_indexeddb_available(self):
+        resp = self.request_lua("""
+        function main(splash)
+            assert(splash:go(splash.args.url))
+            return assert(splash:wait_for_resume(splash.args.js)) 
+        end        
+        """, {'url': self.mockurl('jsrender'), 'js': self.DETECT_INDEXEDDB_JS})
+        self.assertStatusCode(resp, 200)
+        self.assertEqual(resp.json(), {'value': 'ok'})
+
+    def test_indexeddb_enable_disable(self):
+        resp = self.request_lua("""
+        treat = require('treat')
+        function main(splash)
+            splash.indexeddb_enabled = false
+            assert(splash:go(splash.args.url))
+            local res1 = assert(splash:wait_for_resume(splash.args.js))
+            local val1 = splash.indexeddb_enabled
+            
+            splash.indexeddb_enabled = true
+            assert(splash:go(splash.args.url))
+            local res2 = assert(splash:wait_for_resume(splash.args.js))
+            local val2 = splash.indexeddb_enabled
+            return treat.as_array({res1, val1, res2, val2})
+        end        
+        """, {'url': self.mockurl('jsrender'), 'js': self.DETECT_INDEXEDDB_JS})
+        self.assertStatusCode(resp, 200)
+        res1, val1, res2, val2 = resp.json()
+        assert res1 == {'value': 'no indexed DB'}
+        assert val1 is False
+
+        assert res2 == {'value': 'ok'}
+        assert val2 is True
+
+
 class EnableDisablePrivateModeTest(BaseLuaRenderTest):
     LOCAL_STORAGE_WORKS_JS = """
     (function () {